move azure_blob specific code out of azure_common
