+++ src/sinks/azure_blob/mod.rs
@@ -1,5 +1,7 @@
 mod config;
 mod request_builder;
+mod service;
+mod sink;
 
 #[cfg(all(test, feature = "azure-blob-integration-tests"))]
 mod integration_tests;
