+++ src/sinks/azure_blob/sink.rs
