+++ src/sinks/azure_common/mod.rs
@@ -1,5 +1,3 @@
 pub mod config;
 pub mod connection_string;
-pub mod service;
 pub mod shared_key_policy;
-pub mod sink;
